Align `machine_learning`, `metrics` and `utils` with scikit-learn
Fixes a class of divergences where the crate ran without error but produced
different numbers than the scikit-learn pipeline it is meant to replace, which
is the worst failure mode for a library whose stated goal is being a drop-in
replacement. Every change here is verified numerically against a real
scikit-learn 1.9.0 install rather than against recollection: `LinearRegression`
now matches `LinearRegression()` to ~1e-15, `MeanShift` reproduces its cluster
centres, numbering and every label exactly, `estimate_bandwidth` agrees to
1e-14, and `precision_recall_curve` and `roc_curve` match element for element.

Four defects produced wrong numbers silently. `r2_score` and
`explained_variance_score` guarded the undefined constant-target case with an
absolute `1e-10` threshold on an unnormalized sum of squares, so any target
whose entire spread fell below it scored a false `1.0`; constancy is now decided
from the values themselves, since an exact-zero test would fail in the other
direction (`sum / n` does not round-trip every constant).
`precision_recall_curve` appended its closing point at the high-recall end,
leaving recall monotone in neither direction. `KMeans` left `labels_` and
`inertia_` describing the previous iteration's centroids when a fit stopped at
`max_iter`, so `predict(x)` disagreed with `get_labels()`. `MeanShift` weighted
its window with a Gaussian kernel over every sample and merged modes by running
average in seed order, so its centres and cluster count differed from
scikit-learn's throughout.

The remaining changes bring behaviour into line: `LinearRegression` defaults to
the exact closed form; `RegularizationType::L1` applies a proximal step so
penalized coefficients reach exactly `0.0` instead of only approaching it, and
its docs carry the alpha conversion table for each scikit-learn estimator;
`ConfusionMatrix` requires hard labels instead of thresholding both arguments at
`0.5`, with `new_with_labels` for other label pairs; `IsolationForest` adopts
scikit-learn's score sign and gains `decision_function`; `LDA::transform`
centres by the training mean and keeps the whitened axis scale; and cluster
labels are `isize` everywhere with `-1` for noise, so any clustering estimator's
output now feeds any clustering metric.

Two API changes fall out of that work. The two enums both named `Solver` become
`LeastSquaresSolver` and `DiscriminantSolver` — only LDA's was re-exported, so
`Solver::GradientDescent` after a prelude glob resolved to the wrong type and
failed to compile. `LeastSquaresSolver` also carries the settings only gradient
descent uses, which leaves `LinearRegression::new` taking just `fit_intercept`
and makes `default()` equal to `new(true)`, so the two constructors can no
longer build different algorithms. `KMeans` gains `n_init`, defaulting to ten
restarts because the plain k-means++ seeding here lacks the local trials that
justify scikit-learn's `n_init=1`.

Breaking changes and their migrations are recorded in CHANGELOG.md. The guide
updates for all of this stay in the working tree.

+++ benches/benchmarks/silhouette.rs
@@ -30,8 +30,8 @@ fn pseudo_random_matrix(rows: usize, cols: usize, seed: u64) -> Array2<f64> {
 
 /// Round-robin cluster labels: exactly `k` clusters, each of size ~n/k (no singletons), with
 /// `2 <= k < n` so `validate_clustering_inputs` accepts them.
-fn round_robin_labels(n: usize, k: usize) -> Array1<usize> {
-    Array1::from_shape_fn(n, |i| i % k)
+fn round_robin_labels(n: usize, k: usize) -> Array1<isize> {
+    Array1::from_shape_fn(n, |i| (i % k) as isize)
 }
 
 fn bench_silhouette(c: &mut Criterion) {
